Added menu on right click to quit the app
In onRightClick, the menu is assigned, a fake click is performed on the
button, and then the menu is set to nil. This is the only way I found
to display the menu only on right click without using deprecated
methods. It is recommended to call [NSApp terminate:] in this way so
that it happens after the run loop is complete.